backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pdfminer.six
opencv-python
scikit-image
numpy
pyhanko
cryptography
pydantic
google-cloud-aiplatform
google-cloud-storage
python-dotenv

torch
transformers
pillow
